[FEAT]: 시나리오 분석 파일 정리
