+++ src/app/components/scenarios/Analysis.tsx
@@ -3,7 +3,7 @@
 import { PiWarningCircleFill } from "react-icons/pi";
 import { FaLightbulb } from "react-icons/fa";
 import { Tooltip } from "@/share/components/Tooltip";
-import { AnalysisProps } from "../types";
+import { AnalysisProps } from "../../../domains/scenarios/types";
 
 export const Analysis = ({ data }: AnalysisProps) => {
   return (
